chore(merchant): restructure package
