+++ libs/community/google/merchant/garf_merchant_api/exceptions.py
@@ -1,4 +1,4 @@
-# Copyright 2025 Google LLC
+# Copyright 2026 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,8 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from garf_core import exceptions
+import warnings
 
+from garf.community.google.merchant.exceptions import *
 
-class GarfMerchantApiError(exceptions.GarfError):
-  """Base class for all library exceptions."""
+warnings.warn(
+  "The 'garf_merchant_api' namespace is deprecated. "
+  "Please use 'garf.community.google.merchant' instead.",
+  DeprecationWarning,
+  stacklevel=2,
+)
