+++ libs/community/google/merchant/garf_merchant_api/query_editor.py
@@ -1,4 +1,4 @@
-# Copyright 2025 Google LLC
+# Copyright 2026 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,35 +11,14 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Defines MerchantQuery."""
 
-from garf_core import query_editor
-from typing_extensions import Self
+import warnings
 
+from garf.community.google.merchant.query_editor import *
 
-def _to_camel_case(field: str) -> str:
-  first, *others = field.split('_')
-  return ''.join([first.lower(), *map(str.title, others)])
-
-
-class MerchantApiQuery(query_editor.QuerySpecification):
-  """Query to Knowledge Graph Search api."""
-
-  def __init__(
-    self,
-    text: str,
-    title: str | None = None,
-    args: dict[str, str] | None = None,
-    **kwargs,
-  ) -> None:
-    """Initializes MerchantApiQuery."""
-    super().__init__(text, title, args, **kwargs)
-
-  def extract_fields(self) -> Self:
-    for line in self._extract_query_lines():
-      line_elements = query_editor.ExtractedLineElements.from_query_line(line)
-      if field := line_elements.field:
-        self.query.fields.append(
-          _to_camel_case(field) if '_' in field else field
-        )
-    return self
+warnings.warn(
+  "The 'garf_merchant_api' namespace is deprecated. "
+  "Please use 'garf.community.google.merchant' instead.",
+  DeprecationWarning,
+  stacklevel=2,
+)
